IDE-2823 Layout Template
